* src/ne_uri.c: Save a few bytes by using shorts for uri_chars lookup table.

* src/ne_uri.c: Note that only one bit is left for a 16-bit short.

* src/ne_uri.c: Fix comment-in-comment.

+++ neon/branches/0.30.x/src/ne_uri.c
@@ -66,6 +66,7 @@
                      * ... except "+" which is PS */
 
 #define OT (0x4000) /* others */
+/* UNUSED  (0x8000) .. only remaining bit. */
 
 #define URI_ALPHA (AL)
 #define URI_DIGIT (DG)
@@ -92,7 +93,7 @@
 /* any characters which should be path-escaped: */
 #define URI_ESCAPE ((URI_GENDELIM & ~(FS)) | URI_SUBDELIM | OT | PC)
 
-static const unsigned int uri_chars[256] = {
+static const unsigned short uri_chars[256] = {
 /* 0xXX    x0      x2      x4      x6      x8      xA      xC      xE     */
 /*   0x */ OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT,
 /*   1x */ OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT, OT,
